Use argparse, no more need for `--` to pass git options
The main reason for this is that argparse has a more convenient
fall-through mode (parse_known_args) that allows all unknown arguments
to be passed through to the next command. This simplifies the command
syntax in that the `--` before the git log options is not needed
anymore.

A minor side-effect is also that this makes the "crude installation"
easier (i.e. dropping the script into your ~/.local/bin), by removing
the least common dependency: docopt.

(The others are relatively common for most people who do use python, I
assume.)